requirements.txt
streamlit
pandas
openpyxl
pypdf
Pillow
qrcode